[ui] Turn workspace store into a module
Splits the Vuex workspace store into a reusable module.
