+++ ui/src/store/index.js
@@ -1,24 +1,20 @@
 import { createStore } from "vuex";
 import Cookies from "js-cookie";
 import router from "../router";
+import workspaceStore from "./workspace";
 
 export const store = createStore({
   state: {
     user: Cookies.get("sh_user"),
-    workspace: JSON.parse(localStorage.getItem("sh_workspace")) || [],
   },
   mutations: {
     loginUser(state, user) {
       state.user = user;
     },
-    setWorkspace(state, workspaceData) {
-      state.workspace = workspaceData;
-    },
   },
   getters: {
     isAuthenticated: (state) => !!state.user,
     user: (state) => state.user,
-    workspace: (state) => state.workspace,
   },
   actions: {
     async login({ commit }, { username, password }) {
@@ -59,26 +55,8 @@ export const store = createStore({
       Cookies.remove("sh_user");
       router.push({ name: "Login", query: redirect });
     },
-    saveWorkspace({ commit }, workspaceData) {
-      const uuids = workspaceData.map((individual) => individual.uuid);
-      localStorage.setItem("sh_workspace", JSON.stringify(uuids));
-      commit("setWorkspace", uuids);
-    },
-    emptyWorkspace({ commit }) {
-      localStorage.setItem("sh_workspace", JSON.stringify([]));
-      commit("setWorkspace", []);
-    },
-    togglePin({ commit, state }, uuid) {
-      const index = state.workspace.indexOf(uuid);
-      const workspace = state.workspace;
-      if (index === -1) {
-        workspace.push(uuid);
-      } else {
-        workspace.splice(index, 1);
-      }
-      commit("setWorkspace", workspace);
-      localStorage.setItem("sh_workspace", JSON.stringify(workspace));
-    },
   },
-  modules: {},
+  modules: {
+    workspace: workspaceStore,
+  },
 });
